Extract chat session forking helpers
